General fixes to grid files
- Added aiortc to requirements.txt
- Updated bob and maria notebooks to explain how to run network.py
- Moved executable node examples into folder
- Changed executable examples to absolute imports
- Changed grid library code to relative imports
