+++ requirements.txt
@@ -15,6 +15,7 @@
 # numpy==1.13.3
 # scipy==1.0
 #
+aiortc
 bandit
 black
 dataclasses # backport to python 3.6
