+++ src/syft/grid/example_nodes/device.py
@@ -2,11 +2,12 @@
 functionality on an actual local network. This is NOT meant to be run in
 production (that's the *actual* grid's job)."""
 
+
 # third party
 from flask import Flask
 
-# syft relative
-from ..core.node.device.device import Device
+# syft absolute
+from syft.core.node.device.device import Device
 
 app = Flask(__name__)
 device = Device(name="cpu1")
@@ -14,3 +15,6 @@ device = Device(name="cpu1")
 
 def run() -> None:
     app.run()
+
+
+run()
