+++ src/syft/grid/example_nodes/network.py
